pyproject.toml
[tool.poetry]
name = "coursepaper-7"
version = "0.1.0"
description = ""
authors = ["Zhorik Zeniuk <Zhorik.me@list.ru>"]
readme = "README.md"
packages = [{include = "coursepaper_7"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.7"
pillow = "^10.0.1"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"
ipython = "^8.15.0"
celery = "^5.3.4"
eventlet = "^0.33.3"
django-celery-beat = "^2.5.0"
redis = "^5.0.0"
coverage = "^7.3.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
